pyproject.toml
[tool.poetry]
name = "pet_project"
version = "0.1"
description = "pet_project"
authors = ["Alex Kondratenko"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.12.0"
annotated-types = "0.5.0,"
anyio = "3.7.1,"
asyncpg = "0.28.0,"
click = "8.1.7,"
exceptiongroup = "1.1.3,"
fastapi = "0.103.1,"
greenlet = "2.0.2,"
idna = "3.4,"
Mako = "1.2.4"
MarkupSafe = "2.1.3,"
psycopg2-binary = "2.9.7,"
pydantic-core = "2.6.3,"
python-dotenv = "1.0.0,"
sniffio = "1.3.0,"
SQLAlchemy = "2.0.0,"
starlette = "0.27.0,"
typing-extensions = "4.7.1,"
uvicorn = "0.23.2"
h11 = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
